+++ Source/SwiftLintFramework/Configuration+CommandLine.swift
@@ -288,11 +288,12 @@ extension Configuration {
 
     init(options: LintOrAnalyzeOptions) {
         self.init(
-            configurationFiles: options.configurationFiles,
+            configurationFiles: options.effectiveConfigurationFiles,
             enableAllRules: options.enableAllRules,
             onlyRule: options.onlyRule,
             cachePath: options.cachePath
         )
+        basedOnCustomConfigurationFiles = options.configurationFiles.isNotEmpty
     }
 }